Started splitting files in utils
