+++ lib/msal-core/src/AuthResponse.ts
@@ -32,3 +32,15 @@ export function buildResponseStateOnly(state: string) : AuthResponse {
         accountState: state
     };
 }
+
+export function setResponseIdToken(originalResponse: AuthResponse, idToken: IdToken) : AuthResponse {
+    var response = { ...originalResponse };
+    response.idToken = idToken;
+    if (response.idToken.objectId) {
+      response.uniqueId = response.idToken.objectId;
+    } else {
+      response.uniqueId = response.idToken.subject;
+    }
+    response.tenantId = response.idToken.tenantId;
+    return response;
+  }
